photopack/main_system/requirements.txt
pillow==10.4.0
paramiko==3.1.0

